few fixes

+++ src/main/java/ru/rutmiit/user_survey_api/service/implementation/QuestionServiceImpl.java
@@ -4,11 +4,10 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.rutmiit.user_survey_api.exception.QuestionNotFoundException;
-import ru.rutmiit.user_survey_api.exception.SurveyNotFoundException;
 import ru.rutmiit.user_survey_api.model.Question;
 import ru.rutmiit.user_survey_api.repository.QuestionRepository;
-import ru.rutmiit.user_survey_api.repository.SurveyRepository;
 import ru.rutmiit.user_survey_api.service.QuestionService;
+import ru.rutmiit.user_survey_api.service.SurveyService;
 
 import java.util.List;
 
@@ -19,7 +18,7 @@ import static ru.rutmiit.user_survey_api.mapper.ReflectionFieldMapper.mapNonNull
 @RequiredArgsConstructor
 public class QuestionServiceImpl implements QuestionService {
     private final QuestionRepository questionRepository;
-    private final SurveyRepository surveyRepository;
+    private final SurveyService surveyService;
 
     @Override
     public Question findById(Long id) {
@@ -35,8 +34,7 @@ public class QuestionServiceImpl implements QuestionService {
     @Override
     @Transactional
     public Question create(Question question, Long surveyId) {
-        question.setSurvey(surveyRepository.findById(surveyId)
-                .orElseThrow(SurveyNotFoundException::new));
+        question.setSurvey(surveyService.findById(surveyId));
 
         return questionRepository.save(question);
     }
@@ -47,13 +45,6 @@ public class QuestionServiceImpl implements QuestionService {
         var foundedQuestion = this.findById(id);
 
         mapNonNullFields(question, foundedQuestion);
-        var options = question.getOptions();
-        var foundedOptions = foundedQuestion.getOptions();
-        for (int i = 0; i < options.size(); i++){
-            var option = options.get(i);
-            var foundedOption = foundedOptions.get(i);
-            mapNonNullFields(option, foundedOption);
-        }
 
         return questionRepository.save(foundedQuestion);
     }
@@ -61,6 +52,8 @@ public class QuestionServiceImpl implements QuestionService {
     @Override
     @Transactional
     public void deleteById(Long id) {
-        questionRepository.deleteById(id);
+        if (questionRepository.existsById(id))
+            questionRepository.deleteById(id);
+        else throw new QuestionNotFoundException();
     }
 }
